@BewareMyPower 2021-07-27
In this PR:

> kop tests / build (pull_request) Successful in 21m

However, it took 36 minutes in https://github.com/streamnative/kop/runs/3154656221